Crate manganis_core

Modules

hash

Utilities for creating hashed paths to assets in Manganis. This module defines AssetHash which is used to create a hashed path to an asset in both the CLI and the macro.

linker

Utilities for working with Manganis assets in the linker. This module defines LinkSection which has information about what section manganis assets are stored in on each platform.

Structs

Asset

A bundled asset with some options. The asset can be used in rsx! to reference the asset. It should not be read directly with std::fs::read because the path needs to be resolved relative to the bundle

BundledAsset

An asset that should be copied by the bundler with some options. This type will be serialized into the binary and added to the link section LinkSection::CURRENT. CLIs that support manganis, should pull out the assets from the link section, optimize, and write them to the filesystem at BundledAsset::bundled_path for the application to use.

CssAssetOptions

Options for a css asset

FolderAssetOptions

The builder for [FolderAsset]

ImageAssetOptions

Options for an image asset

JsAssetOptions

Options for a javascript asset

Enums

AssetOptions

Settings for a generic asset

ImageFormat

The type of an image. You can read more about the tradeoffs between image formats here

ImageSize

The size of an image asset